Ajout de l'extension pour la jauge
