+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "bootstrap": "^5.3.0-alpha1",
+    "jquery": "^3.6.3",
     "mdb-react-ui-kit": "^5.1.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.2",
@@ -18,6 +19,7 @@
     "react-toastify": "^9.1.1"
   },
   "devDependencies": {
+    "@types/jquery": "^3.5.16",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@vitejs/plugin-react": "^3.1.0",
