--- src/accueil.css
@@ -1,138 +1,0 @@
-
-#html {
-  background-color: #f5f5f5;
-}
-
-.Content{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-  background-color: #f5f5f5;
-  width: 100%;
-  margin-top: -20%;
-}
-
-.logoContent {    
-  padding: 0 1.5em;
-}
-
-.logoContent img {
-  width: 90% 
-}
-
-.logoSection1 {
-  display: flex;
-  justify-content: space-around;
-  height: 40%;
-  max-width: 42%;
-  margin-bottom: -35%;
-  margin-right: 50%;
-}
-
-.logoSection2 {
-  display: flex;
-  justify-content: space-around;
-  height: 40%;
-  max-width: 42%;
-  margin-bottom: -10%;
-  margin-left: 50%;
-}
-
-.logoSection1 button {
-  max-height: 50%;
-  color: #002d45;;
-}
-
-.logoSection2 button {
-  max-height: 50%;
-  color: #002d45;
-}
-
-.logoGeneralites {
-  max-width: 100%;
-  
-}
-
-.logoRendezVous {
-  max-width: 100%;
-}
-
-.logoAutoEvaluation {
-  max-width: 50%;
-  background-color: #f5f5f5;
-}
-
-.logoRessource {
-  max-width: 60%;
-}
-
-@font-face {
-  font-family: amaranth;
-  src: url(./assets/polices/amaranth/Amaranth-Regular.otf);
-}
-
-hr {
-  margin-top: 3em;
-  width: 65em;
-}
-
-button {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-top: 1.5em;
-  font-family: amaranth;
-  font-weight: bold !important;
-  background-color: #f5f5f5 !important; 
-}
-
-button:hover {
-  color: #002d45 !important;
-}
-
-@media (min-width: 640px) {
-  .textLogin {
-    font-size: 300%;
-  }
-  .textInscription {
-    font-size: 200%;
-  }
-  .logoContent img {
-    max-width: 300px;
-  }
-  .choixContent {
-    font-size: 200%;
-  }
-  .Content {
-    width: 100%;
-  }
-  .logoContent {
-    padding: 0 2em;
-  }
-  hr {
-    width: 40em;
-    margin-top: 1em;
-  }
-  input {
-    width: 35em;
-  }
-  button {
-    margin-top: 2em;
-  }
-}
-
-.btn-secondary{
-  background-color: #f5f5f5 !important;
-  border-color: #002d45 !important;
-}
-
-.text-primary{
-  color:#002d45!important;
-}
-
-
-.btn-exit{
-  font-size: 1.8em;
-  color:#002d45;
-}
