+++ src/extensions/jauge/jauge.css
@@ -1,0 +1,46 @@
+.cluster{
+    position: relative;
+    width: 50%;
+    bottom: 20px;
+    z-index: 2;
+    transform: rotate(180deg);
+    overflow: hidden;
+
+}
+.clusterLine{
+    width: 50%;
+    position: relative;
+    border-bottom: black solid 3px;
+}
+
+input[type=range] {
+    -webkit-appearance: none;
+    width: 100%;
+    height: 40px;
+    background: #d3d3d3;
+    outline: none;
+    opacity: 0.7;
+    -webkit-transition: .2s;
+    transition: opacity .2s;
+}
+
+input[type=range]:hover {
+    opacity: 1;
+}
+
+input[type=range]::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    appearance: none;
+    width: 40px;
+    height: 40px;
+    background: #ea3c3c;
+    cursor: pointer;
+}
+
+input[type=range]::-moz-range-thumb {
+    width: 40px;
+    height: 40px;
+    background: #ea3c3c;
+    cursor: pointer;
+}
+
