+++ src/index.css
@@ -14,7 +14,11 @@ html, body {margin: 0; height: 100%; overflow: hidden}
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   -webkit-text-size-adjust: 100%;
-  background-color: #f5f5f5;
+}
+
+@font-face {
+  font-family: amaranth;
+  src: url(./assets/polices/amaranth/Amaranth-Regular.otf);
 }
 
 a {
