+++ src/page/generalites/generalite.css
@@ -1,0 +1,27 @@
+.jauge{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+}
+
+.jauge img{
+    width: 75%;
+    height: 100%;
+}
+
+.cluster{
+    position: relative;
+    width: 50%;
+    bottom: 20px;
+    z-index: 2;
+    transform: rotate(180deg);
+    overflow: hidden;
+
+}
+.clusterLine{
+    width: 50%;
+    position: relative;
+    border-bottom: black solid 3px;
+}
+
