+++ src/page/login/login.css
@@ -1,0 +1,151 @@
+
+.Content{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: #f5f5f5;
+    
+    background-repeat: no-repeat;
+    background-size: cover;
+    position: relative;
+    z-index: 1;
+    height: 370vw;
+    padding-top:5em;
+    
+}
+
+.backgroundContent, .backgroundContent2{
+    position: absolute;
+    z-index: -1;
+}
+.backgroundContent{
+    top: 140vw;
+}
+
+
+.backgroundContent img {
+    width: 100%;
+    height: 100%;
+    opacity: 0.2;
+}
+
+.logoContent {    
+    margin-top: 1em;
+    padding: 0 1.5em;
+    display: flex;
+    justify-content: center;
+}
+
+.logoContent img {
+    width: 90%;
+}
+
+
+.textLogin {
+    font-size: 250%;
+    color: black;
+    font-family: amaranth;
+    text-align: center;
+}
+
+.textIncription {
+    font-size: 100%;
+    color: black;
+    font-family: amaranth;
+    text-align: center;
+}
+
+
+
+hr {
+    margin-top: 3em;
+    width: 65em;
+}
+
+.inputContent{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+
+input {
+    margin-top: 0.5em !important;
+    width: 100%;
+    height:1em;
+    border-radius: 5px;
+    padding: 0.5em 1em;
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    border: 1px solid #002D45;
+    background-color: #f5f5f5;
+    transition: border-color 0.25s;
+    color: black;
+}
+
+input[type="submit"]{
+    color: black;
+    width: 10em;
+    height: 2.5em;
+    border-radius: 5px;
+    border: 1px solid black;
+    margin-top: 2em;
+}
+
+input:focus, input:focus-visible {
+    outline: 4px auto -webkit-focus-ring-color;
+}
+
+button {
+    display: flex;
+    align-items: center;
+    margin-top: 1.5em;
+    font-family: amaranth;
+    background-color: #002D45; 
+    color: white;
+}
+
+.buttonContent {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.buttonContentLogin{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.buttonContentLogin button{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #213547 !important; 
+}
+.choixContent {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 1em;
+    font-size: 50%;
+    color: #002D45;
+}
+
+.inscriptionContent{
+    display:flex;
+    justify-content:center;
+    flex-direction:column;
+    align-items:center;
+    z-index:1;
+}
+
+.backgroundLogin{
+    position: absolute;
+    width: 100%;
+    opacity: 0.2;
+    z-index: -1;
+}
+
+.btn-primary{
+    background-color: #002D45 !important;
+}
